requirements.txt
Flask>=2.0
python-nmap>=0.6.1
